fix animation on home page
because this is all happening client side,
changing the page stops the animation running
in the background on the home page. navigating
back doesnt restart it since the animation
only expected to start itself once.

this uses componentDidMount to check whenever
the home page has been reloaded, to send a message
off to the JS animation (constellation.js)

Instead of creating the canvas in the
render function, this creates it in JS,
 so the animation can be sure it exists

have to keep the script files in the layout
and not in the index file, so that they
dont get diff'ed out of the DOM,
and the window.initPlugin function becomes
undefined.